base/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
watchdog==3.0.0
filelock==3.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0,<4.1.0
markdown==3.5.1
pygments==2.17.2
jinja2==3.1.2
aiohttp==3.9.1
